+++ crates/wasm-pkg-core/src/resolver.rs
@@ -2,7 +2,7 @@
 // NOTE(thomastaylor312): This is copied and adapted from the `cargo-component` crate: https://github.com/bytecodealliance/cargo-component/blob/f0be1c7d9917aa97e9102e69e3b838dae38d624b/crates/core/src/registry.rs
 
 use std::{
-    collections::{hash_map, HashMap, HashSet},
+    collections::{hash_map, BTreeSet, HashMap, HashSet},
     fmt::Debug,
     ops::{Deref, DerefMut},
     path::{Path, PathBuf},
@@ -14,7 +14,7 @@ use futures_util::TryStreamExt;
 use indexmap::{IndexMap, IndexSet};
 use petgraph::{
     acyclic::{Acyclic, AcyclicEdgeError},
-    graph::DiGraph,
+    graph::{DiGraph, NodeIndex},
     Graph,
 };
 use semver::{Comparator, Op, Version, VersionReq};
@@ -23,11 +23,14 @@ use wasm_pkg_client::{
     caching::{CachingClient, FileCache},
     Client, Config, ContentDigest, Error as WasmPkgError, PackageRef, Release, VersionInfo,
 };
-use wasm_pkg_common::registry::DependencyOf;
+use wasm_pkg_common::registry::{DependencyGraph, DependencyOf};
 use wit_component::DecodedWasm;
 use wit_parser::{PackageId, PackageName, Resolve, UnresolvedPackageGroup, WorldId};
 
-use crate::{lock::LockFile, wit::get_packages};
+use crate::{
+    lock::LockFile,
+    wit::{get_local_dependencies, get_packages},
+};
 
 /// The name of the default registry.
 pub const DEFAULT_REGISTRY_NAME: &str = "default";
@@ -857,4 +860,91 @@ fn visit<'a>(
     assert!(order.insert(dep.package_name().clone()));
 
     Ok(())
+}
+
+/// State for tracking dependencies during upload.
+pub struct PublishPlan {
+    /// Graph of publishable packages where the edges are `(dependency -DependencyOf->) dependent)`
+    dependents: DependencyGraph<PackageRef>,
+    /// Mapping [`PackageRef`]s to the respective index inside the dependency graph.
+    // TODO look at using cargo's `InternedString` type for `PackageRef`:
+    // https://docs.rs/cargo/latest/cargo/util/interning/struct.InternedString.html
+    indices: HashMap<PackageRef, NodeIndex>,
+}
+
+impl PublishPlan {
+    /// Generate [`Self`] from a list of WIT package paths (files or directories).
+    fn from_paths(paths: &[impl AsRef<Path>]) -> Result<Self> {
+        let (graph, indices) = get_local_dependencies(paths)?;
+        let mut dependents = graph.into_inner();
+        dependents.reverse();
+        // graph was already found to be acyclic
+        let dependents = DependencyGraph::try_from(dependents).unwrap();
+
+        Ok(Self {
+            dependents,
+            indices,
+        })
+    }
+
+    fn iter<'a>(&'a self) -> impl Iterator<Item = &'a PackageRef> + 'a {
+        self.indices.iter().map(|(pkg, _)| pkg)
+    }
+
+    fn is_empty(&self) -> bool {
+        self.indices.is_empty()
+    }
+
+    fn len(&self) -> usize {
+        self.indices.len()
+    }
+
+    /// Returns the set of packages that are ready for publishing (i.e. have no outstanding dependencies).
+    ///
+    /// These will not be returned in future calls.
+    fn take_ready(&mut self) -> BTreeSet<PackageRef> {
+        self.dependents
+            .nodes_iter()
+            // there are no dependents on `self.dendents[id]`
+            .filter(|id| self.dependents.neighbors(*id).count() == 0)
+            .map(|id| {
+                let pkg = &self.dependents[id];
+                self.indices.remove(&pkg);
+                pkg.clone()
+            })
+            .collect()
+    }
+
+    /// Packages confirmed to be available in the registry, potentially allowing additional
+    /// packages to be "ready".
+    fn mark_confirmed(&mut self, published: impl IntoIterator<Item = PackageRef>) {
+        for pkg in published {
+            let id = self
+                .indices
+                .remove(&pkg)
+                .expect("PackageRef has no associated index");
+            self.dependents
+                .remove_node(id)
+                .expect("index has no associated PackageRef");
+        }
+    }
+}
+
+/// Format a collection of packages as a list
+///
+/// e.g. "foo:a@0.1.0, bar:b@0.2.0, and baz:c@0.3.0".
+///
+/// Note: the final separator (e.g. "and" in the previous example) can be chosen.
+fn package_list(pkgs: impl IntoIterator<Item = PackageRef>, final_sep: &str) -> String {
+    let mut names: Vec<_> = pkgs.into_iter().map(|pkg| pkg.to_string()).collect();
+    names.sort();
+
+    match &names[..] {
+        [] => String::new(),
+        [a] => a.clone(),
+        [a, b] => format!("{a} {final_sep} {b}"),
+        [names @ .., last] => {
+            format!("{}, {final_sep} {last}", names.join(", "))
+        }
+    }
 }
